mcps/vectorworks-mcp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
httpx==0.27.2
websockets==12.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
orjson==3.10.7
python-dotenv==1.0.1
